Rename assert_hw to assertion
