+++ transactron/utils/__init__.py
@@ -4,6 +4,6 @@ from .debug_signals import *  # noqa: F401
 from .assign import *  # noqa: F401
 from .amaranth_ext import *  # noqa: F401
 from .transactron_helpers import *  # noqa: F401
-from .assert_hw import *  # noqa: F401
+from .assertion import *  # noqa: F401
 from .dependencies import *  # noqa: F401
 from .depcache import *  # noqa: F401
